fix: zio-json-macros for Scala 3
Enabled Scala 3 build for zio-json-macros by splitting sources/tests by Scala version: kept full Scala 2 macro-annotation behavior, added a Scala 3 source-compatible @jsonDerive annotation plus Scala
