+++ zio-json-macros/shared/src/main/scala/zio/json/JsonDeriveConfig.scala
@@ -1,0 +1,17 @@
+package zio.json
+
+/**
+ * Define a config for derivation macro
+ */
+sealed abstract class JsonDeriveConfig
+
+object JsonDeriveConfig {
+  // Derive a JsonCodec
+  case object Codec extends JsonDeriveConfig
+
+  // Derive only a JsonEncoder
+  case object Encoder extends JsonDeriveConfig
+
+  // Derive only a JsonDecoder
+  case object Decoder extends JsonDeriveConfig
+}
